+++ __tests__/index.spec.ts
@@ -1,73 +1,77 @@
 import axios, { AxiosInstance } from 'axios'
-import mockServer, { MockModels, MockRouter, Datastore } from '~/src'
+import fs from 'fs'
+import path from 'path'
+import mockServer, { MockRouter, DataStore, Seeds, asyncResponse } from '~/src'
+import formToBuffer from './libs/formToBuffer'
 
 describe('initialize', () => {
   let client: AxiosInstance
-  let db: Datastore
+  let dataStore: DataStore
 
   beforeEach(() => {
-    client = axios.create({
-      baseURL: 'https://google.com'
-    })
+    client = axios.create({ baseURL: 'https://google.com' })
+    dataStore = new DataStore()
   })
 
-  afterEach(() => db.deleteAll())
+  afterEach(() => dataStore.deleteAll())
 
   test('enabled mock', async () => {
-    const models = {}
-    const router:any[] = []
+    const router: MockRouter = []
 
-    db = (await mockServer(client, models, router)).db
+    await mockServer(client, router)
     await expect(client.get('/')).rejects.toHaveProperty('response.status', 404)
   })
 
   test('get', async () => {
     const testPath = '/test'
     const defaultValue = { name: 'test' }
-    const models: MockModels = {
-      test: {
-        seeds: () => [defaultValue],
-        create: (db, { name }) => ({ name })
-      }
-    }
-
+    const seeds: Seeds = { test: [defaultValue] }
     const router: MockRouter = [
       {
         path: testPath,
         methods: {
-          get: (db) => db.getCollection('test').asyncFind({})
+          get: () => asyncResponse(200, dataStore.getCollection('test').asyncFind({}))
         }
       }
     ]
 
-    db = (await mockServer(client, models, router)).db
+    dataStore.initCollectionsAndSeeds(seeds)
+    await mockServer(client, router)
     const mockedData = await client.get(testPath)
 
     expect(mockedData.data).toHaveLength(1)
     expect(mockedData.data[0]).toEqual({ ...defaultValue, _id: expect.any(String) })
   })
 
+  test('404 request', async () => {
+    const testPath = '/test'
+    const router: MockRouter = [
+      {
+        path: testPath,
+        methods: {}
+      }
+    ]
+
+    await mockServer(client, router)
+    await expect(client.get(testPath)).rejects.toHaveProperty('response.status', 404)
+  })
+
   test('get with params', async () => {
     const testRegPath = '/test/_name'
     const testPath = '/test/sample'
     const defaultValue = { name: 'sample' }
-    const models: MockModels = {
-      test: {
-        seeds: () => [defaultValue, { name: 'sample2' }],
-        create: (db, { name }) => ({ name })
-      }
-    }
-
+    const seeds: Seeds = { test: [defaultValue, { name: 'sample2' }] }
     const router: MockRouter = [
       {
         path: testRegPath,
         methods: {
-          get: (db, params) => db.getCollection('test').asyncFind(params)
+          get: ({ params }) => asyncResponse(200, dataStore.getCollection('test').asyncFind(params))
         }
       }
     ]
 
-    db = (await mockServer(client, models, router)).db
+    dataStore.initCollectionsAndSeeds(seeds)
+    await mockServer(client, router)
     const mockedData = await client.get(testPath)
 
     expect(mockedData.data).toHaveLength(1)
@@ -78,26 +82,21 @@ describe('initialize', () => {
     const testRegPath = '/test/_name'
     const testPath = '/test/sample2'
     const defaultValue = { name: 'sample2', title: 'bbb' }
-    const models: MockModels = {
-      test: {
-        seeds: () => [{ name: 'sample', title: 'aaa' }],
-        create: (db, { name, title }) => ({ name, title })
-      }
-    }
-
+    const seeds: Seeds = { test: [{ name: 'sample', title: 'aaa' }] }
     const router: MockRouter = [
       {
         path: testRegPath,
         methods: {
-          get: (db, params) => db.getCollection('test').asyncFind(params),
-          post: (db, { name }, { title }) => db.getCollection('test').asyncInsert(
-            models.test.create(db, { name, title })
-          )
+          get: ({ params }) =>
+            asyncResponse(200, dataStore.getCollection('test').asyncFind(params)),
+          post: ({ params: { name }, data: { title } }) =>
+            asyncResponse(201, dataStore.getCollection('test').asyncInsert({ name, title }))
         }
       }
     ]
 
-    db = (await mockServer(client, models, router)).db
+    dataStore.initCollectionsAndSeeds(seeds)
+    await mockServer(client, router)
     const mockedData = await client.post(testPath, { title: 'bbb' })
 
     expect(mockedData.data).toEqual({ ...defaultValue, _id: expect.any(String) })
@@ -105,5 +104,40 @@ describe('initialize', () => {
     const mockedData2 = await client.get(testPath)
     expect(mockedData2.data).toHaveLength(1)
     expect(mockedData2.data[0]).toEqual({ ...defaultValue, _id: expect.any(String) })
+  })
+
+  test('post with multipart/form-data', async () => {
+    const testPath = '/test'
+    const name = 'sample-name'
+    const imageBinary = fs.readFileSync(path.join(__dirname, './assets/logo.png'))
+    const formData: [string, string | number | Buffer][] = [
+      ['name', name],
+      ['num', 123],
+      ['file', imageBinary]
+    ]
+    const imageDataURI = fs.readFileSync(path.join(__dirname, './assets/logo.b64'), 'utf-8')
+    const seeds: Seeds = { test: [] }
+    const router: MockRouter = [
+      {
+        path: testPath,
+        methods: {
+          get: () => asyncResponse(200, dataStore.getCollection('test').asyncFind({})),
+          post: ({ data: { file } }) =>
+            asyncResponse(201, dataStore.getCollection('test').asyncInsert({ file }))
+        }
+      }
+    ]
+
+    dataStore.initCollectionsAndSeeds(seeds)
+    await mockServer(client, router)
+    const { boundary, buffer } = formToBuffer(formData)
+    const mockedData = await client.post(testPath, buffer, {
+      headers: { 'Content-Type': `multipart/form-data; boundary=${boundary}` }
+    })
+    expect(mockedData.data).toEqual({ file: imageDataURI, _id: expect.any(String) })
+
+    const mockedData2 = await client.get(testPath)
+    expect(mockedData2.data).toHaveLength(1)
+    expect(mockedData2.data[0]).toEqual({ file: imageDataURI, _id: expect.any(String) })
   })
 })