+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 0.1.0 (2019-08-21)
+
+
+### Bug Fixes
+
+* test ([7c31ad6](https://github.com/m-mitsuhide/axios-mock-server/commit/7c31ad6))
+* url params ([a5c4631](https://github.com/m-mitsuhide/axios-mock-server/commit/a5c4631))
+
+
+### Features
+
+* **change:** response of HTTP methods ([254c1a9](https://github.com/m-mitsuhide/axios-mock-server/commit/254c1a9))
+* **changing:** separated datastore ([b5bf246](https://github.com/m-mitsuhide/axios-mock-server/commit/b5bf246))
+* **http post:** multipart/form-data ([daaebfa](https://github.com/m-mitsuhide/axios-mock-server/commit/daaebfa))
+
 ### 0.0.1 (2019-08-18)